rename module slab to service
