+++ ext/http/http_next.rs
@@ -10,11 +10,11 @@ use crate::request_properties::HttpPropertyExtractor;
 use crate::response_body::Compression;
 use crate::response_body::ResponseBytes;
 use crate::response_body::ResponseBytesInner;
-use crate::slab::handle_request;
-use crate::slab::http_trace;
-use crate::slab::HttpRecord;
-use crate::slab::HttpRequestBodyAutocloser;
-use crate::slab::RefCount;
+use crate::service::handle_request;
+use crate::service::http_trace;
+use crate::service::HttpRecord;
+use crate::service::HttpRequestBodyAutocloser;
+use crate::service::RefCount;
 use crate::websocket_upgrade::WebSocketUpgrade;
 use crate::LocalExecutor;
 use cache_control::CacheControl;
